+++ CHANGELOG.md
@@ -438,3 +438,4 @@
 - [2026-04-16 17:32] refactor(perf): decouple repository interface from concrete perf service
 - [2026-04-16 18:02] refactor(i18n): decouple repository interface from concrete i18n service
 - [2026-04-16 18:09] feat(test): integrate custom telemetry logger in test
+- [2026-04-16 18:25] refactor(auth): decouple repository interface from concrete auth service